Move DB setup and teardown to each test class
When testing endpoints which for example creates users,
there is no way to rollback the database. Therefore when testing in random orders
as a new user is created and the get all endpoint is triggered there are now not just two test users
but 3.

This solution is a bit slower when executing but probably better to have deterministic tests.

An attempt has been made to use QuarkusTestBeforeEachCallback and
QuarkusTestAfterEachCallback as service providers with no luck as the CDI context was not loaded properly